sps_basis reorganization and cleanup
Moves StellarPopBasis to galaxymodel, and move StarBasis objects to starmodel.  Many cleanups in StellarPopBasis code. Start on a CSPBasis object in galaxymodel that will replace CSPModel in sedmodel.
